made labelling more intelligent

+++ data/pfp_odgt.py
@@ -90,15 +90,21 @@ if __name__ == '__main__':
 
     print(f'Train indexing')
     odgt_fp_train = os.path.join(dataset_dir_p, f'{TRAIN_NAME}_{ODGT_NAME}')
-    open(odgt_fp_train, 'w').close() 
-    indices2odgt(odgt_fp_train, dataset_dir_p, indices[:limit], img_fps, mask_fps)
-    print(f'Train file saved at: {odgt_fp_train}\n')
+    if os.path.exists(odgt_fp_train):
+        print('Train indexing already done!')
+    else:
+        open(odgt_fp_train, 'w').close() 
+        indices2odgt(odgt_fp_train, dataset_dir_p, indices[:limit], img_fps, mask_fps)
+        print(f'Train file saved at: {odgt_fp_train}\n')
 
     print(f'Validation indexing')
     odgt_fp_val = os.path.join(dataset_dir_p, f'{VAL_NAME}_{ODGT_NAME}')
-    open(odgt_fp_val, 'w').close() 
-    indices2odgt(odgt_fp_val, dataset_dir_p, indices[limit:], img_fps, mask_fps)
-    print(f'Validation file saved at: {odgt_fp_val}')
+    if os.path.exists(odgt_fp_val):
+        print('Validation indexing already done!')
+    else:
+        open(odgt_fp_val, 'w').close() 
+        indices2odgt(odgt_fp_val, dataset_dir_p, indices[limit:], img_fps, mask_fps)
+        print(f'Validation file saved at: {odgt_fp_val}')
